+++ solver_rust_conic/src/solver_error.rs
@@ -1,0 +1,4 @@
+use crate::solver::SolverError;
+use std::error::Error;
+
+impl Error for SolverError {}